Move App and test to src folder in template
